Fix unreliable IPv4 internet test in alpine-install.func
Added two additional public resolver IPs (8.8.8.8 & 9.9.9.9) from diverse ASNs so that if any of the three IPs are reachable, the test will succeed.
The only condition that should fail this test is if the configuration is incorrect, and there is no internet access at all... therefore, if even one of the test destinations responds, the container network configuration is correct and the test should succeed.

Three diverse autonomous systems is the minimum number of destinations that should be checked for a connectivity test like this.
More would be better, but three will do.   If ALL fail, the test fails. If ANY respond, the test should succeed. 
The more (and more diverse) destination ASNs checked, the lower the probability that the test returns a false "fail" due to network conditions (such as routing problems on the upstream ISP, packet loss, firewall blocks, etc.) other than incorrect container configuration, which is what this test is meant to validate.