Update filebrowser.sh
The default port 8080 seems to be a popular port for a lot of programs, and while some of them may allow you to change the port when installing, most don't.  I have changed the script to give you a prompt to select a different port than 8080 if you so decide, but allows you to keep the default port if wanted.