Try to catch download errors
During troubleshooting of another error, I ran into a condition where downloading the new version failed. It seems that this is because the earlier call to get the $RELEASE had failed.

This change is an attempt to catch download errors, and alert the user, and to not rm the existing install until we know the download worked.