fix: add documenso again
I reviewed the now removed Documenso script for Proxmox
and saw that it handled installation by attempting to
build from source which while it should work can be prone
to issues.

The issue that triggered the removal is unrelated but is
likely a consequence of `dotenv` having some weirdness
with LXC containers.

The new approach uses a docker container like Komodo and
a few other installers do that simplifies running and
updating the Documenso installation.

NOTE: This is untested, I personally don't run Proxmox
and don't have the time to do so for verification. I've
simply made these changes to assist self-hosters who are
running Proxmox.