fix(pangolin-install): add network-online dependency
Without it, traefik will try to start before the network is ready.
Which, in turn, means that it won't be able to activate the `badger`
plugin. And without that plugin, non of the Pangolin-backed routes in
Traefik will work.

This patch adds a dependency on `network-online.service` for the
generated `traefik.service`, to avoid this issue.